Separate AppHelpers from Component Presenters
This PR is an attempt to structure the app in such a way that we
separate internal classes and helpers from publicly exposed classes.

This hopefully will solve a dilemma we've been having with
govuk_navigation_helpers. That gem was introduced to generate payloads
for components. This was a suboptimal solution to the problem that we
couldn't reasonably add more complex logic to the components in Static
(see the analytics component a cautionary example).

Now that we are moving all components to this gem, we are better able
to use helpers and there's no more need to separate the gems that
create payloads and the components. I plan to merge in
govuk_navigation_helpers in #207.

https://trello.com/c/roHUGIpK
