+++ spec/dummy/app/controllers/step_nav_controller.rb
@@ -1,7 +1,7 @@
 class StepNavController < ApplicationController
   def show
     content_item = Services.content_store.content_item(base_path)
-    @step_nav_helper = GovukPublishingComponents::StepNavHelper.new(content_item, base_path)
+    @step_nav_helper = GovukPublishingComponents::AppHelpers::StepNavHelper.new(content_item, base_path)
   end
 
 private
